VEA-44 Application-wide exception handling.
Moved exception handlers in controllers to exception handlers in an
application-wide controller advice component.
