+++ src/main/java/com/danieldigiovanni/email/auth/AuthController.java
@@ -4,22 +4,12 @@ import com.danieldigiovanni.email.auth.dto.AuthResponse;
 import com.danieldigiovanni.email.auth.dto.LoginRequest;
 import com.danieldigiovanni.email.auth.dto.RegisterRequest;
 import jakarta.persistence.EntityExistsException;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 import jakarta.validation.ValidationException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Map;
-import java.util.stream.Collectors;
 
 @RestController
 public class AuthController {
@@ -39,43 +29,6 @@ public class AuthController {
     @PostMapping("/login")
     public AuthResponse login(@RequestBody @Valid LoginRequest loginRequest) {
         return this.authService.login(loginRequest);
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public Map<String, String> handleValidationException(MethodArgumentNotValidException exception) {
-        return exception.getFieldErrors()
-            .stream()
-            .collect(Collectors.toMap(
-                FieldError::getField,
-                fieldError -> fieldError.getDefaultMessage() == null
-                    ? "Unknown error"
-                    : fieldError.getDefaultMessage()
-            ));
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(ValidationException.class)
-    public String handleValidationException(ValidationException exception) {
-        return exception.getMessage();
-    }
-
-    @ResponseStatus(HttpStatus.CONFLICT)
-    @ExceptionHandler(EntityExistsException.class)
-    public String handleAlreadyExistsException(EntityExistsException exception) {
-        return exception.getMessage();
-    }
-
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(EntityNotFoundException.class)
-    public String handleNotFoundException(EntityNotFoundException exception) {
-        return exception.getMessage();
-    }
-
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    @ExceptionHandler(AuthenticationException.class)
-    public String handleNotFoundException(AuthenticationException exception) {
-        return exception.getMessage();
     }
 
 }
